Include `windows-rs` documentation in the generated docs
None of the predefined types like `Result`, `HRESULT` or any other type
in `windows-rs` are clickable links to the respective documentation.
The `--extern-html-root-url` feature is still unstable but could
otherwise be used to link to external documentation like
`docs.rs/windows`.

Note that by explicitly specifying what crates to document, both the
`windows` and `bindings` crate have to be listed.